feat(transformers): weights initialization now 28x faster on a large scale, added a DecoderGroup to stack multiple transformer decoders at once, new test for big transformers chore(library): changed parameters counter from int to long, optimized imports

+++ brain4j-math/src/main/java/net/echo/math4j/BrainUtils.java
@@ -74,7 +74,7 @@ public class BrainUtils {
         return Math.max(Math.min(gradient, GRADIENT_CLIP), -GRADIENT_CLIP);
     }
 
-    public static String formatNumber(int params) {
+    public static String formatNumber(long params) {
         String[] prefixes = {"B", "KB", "MB", "GB", "TB"};
 
         if (params == 0) {
